rename module files for consistency
